Rename .exc to .utils
