--- src/test_typing_extensions/ext.py
@@ -1,5 +1,0 @@
-# We just import the ext module to check for syntax or import problems.
-import typing_extensions.ext  # noqa F401
-
-def dummy():
-    pass
